Move apis&cache into pkg/controllers folder
