+++ hack/.golint_failures
@@ -7,9 +7,9 @@ volcano.sh/volcano/pkg/apis/batch/v1alpha1
 volcano.sh/volcano/pkg/apis/bus/v1alpha1
 volcano.sh/volcano/pkg/apis/helpers
 volcano.sh/volcano/pkg/cli/job
+volcano.sh/volcano/pkg/controllers/apis
+volcano.sh/volcano/pkg/controllers/cache
 volcano.sh/volcano/pkg/controllers/job
-volcano.sh/volcano/pkg/controllers/job/apis
-volcano.sh/volcano/pkg/controllers/job/cache
 volcano.sh/volcano/pkg/controllers/job/helpers
 volcano.sh/volcano/pkg/controllers/job/plugins
 volcano.sh/volcano/pkg/controllers/job/plugins/env
