+++ pkg/controllers/job/job_controller_handler.go
@@ -29,8 +29,8 @@ import (
 	kbtype "github.com/kubernetes-sigs/kube-batch/pkg/apis/scheduling/v1alpha1"
 	vkbatchv1 "volcano.sh/volcano/pkg/apis/batch/v1alpha1"
 	vkbusv1 "volcano.sh/volcano/pkg/apis/bus/v1alpha1"
-	"volcano.sh/volcano/pkg/controllers/job/apis"
-	vkcache "volcano.sh/volcano/pkg/controllers/job/cache"
+	"volcano.sh/volcano/pkg/controllers/apis"
+	vkcache "volcano.sh/volcano/pkg/controllers/cache"
 )
 
 func (cc *Controller) addCommand(obj interface{}) {
